pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rightmove-scraper"
version = "0.1.0"
description = "Polite Rightmove listing scraper for personal research"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{ name = "Kevin Lee" }]
dependencies = [
  "playwright>=1.45.0",
  "pydantic>=2.6.0",
  "typer[all]>=0.12.0",
  "pandas>=2.2.0",
  "pyarrow>=15.0.0",
  "tenacity>=8.2.0",
  "rich>=13.7.0",
  "python-dotenv>=1.0.1",
  "lxml>=5.2.0",
  "geopy>=2.4.0",
  "snowflake-sqlalchemy>=1.5.0",
  "snowflake-connector-python>=3.7.0",
  "sqlalchemy>=2.0.0",
  "dask>=2024.1.0",
  "numpy>=1.24.0",
  "tqdm>=4.66.0",
  "shapely>=2.0.0",
  "requests>=2.32.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.1.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.5.0",
  "black>=24.8.0",
  "mypy>=1.10.0",
  "pre-commit>=3.8.0",
  "mkdocs-material>=9.5.0"
]

[project.scripts]
rightmove-scraper = "rightmove_scraper.cli:app"

[tool.setuptools.packages.find]
where = ["src"]



[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
lint.select = ["E", "F", "I", "B", "UP"]
lint.ignore = []

[tool.mypy]
python_version = "3.11"
strict = false
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true
